fixed wrong imports and removed duplicities regarding cookeis store
- some code was duplicated, while other was not
necessary at all
- these blocks of code were removed, or moved under
correct folders
