+++ project-base/storefront/components/Pages/App/AppPageContent.tsx
@@ -8,7 +8,7 @@ import { GtmProvider } from 'gtm/context/GtmProvider';
 import { NextComponentType, NextPageContext } from 'next';
 import dynamic from 'next/dynamic';
 import { ToastContainer } from 'react-toastify';
-import { useCookiesStoreSync } from 'utils/cookies/cookieStoreUtils';
+import { useCookiesStoreSync } from 'utils/cookies/cookiesStore';
 import { ServerSidePropsType } from 'utils/serverSide/initServerSideProps';
 
 const Error503Content = dynamic(
