+++ project-base/storefront/store/useCookiesStore.ts
@@ -1,47 +1,7 @@
 import { CookiesStoreContext } from 'components/providers/CookiesStoreProvider';
 import { useContext } from 'react';
-import { v4 as uuidV4 } from 'uuid';
+import { CookiesStore } from 'utils/cookies/cookiesStore';
 import { useStore } from 'zustand';
-import { createStore } from 'zustand/vanilla';
-
-export type CookiesStoreState = {
-    lastVisitedProductsCatnums: string[] | null;
-    userIdentifier: string;
-};
-
-export type CookiesStoreActions = {
-    setCookiesStoreState: (value: Partial<CookiesStoreState>) => void;
-};
-
-export type CookiesStore = CookiesStoreState & CookiesStoreActions;
-
-export const getDefaultInitState = (): CookiesStoreState => ({
-    lastVisitedProductsCatnums: null,
-    userIdentifier: uuidV4(),
-});
-
-export const createCookiesStore = (cookieStoreFromServer: CookiesStoreState) =>
-    createStore<CookiesStore>()((set) => ({
-        ...cookieStoreFromServer,
-        setCookiesStoreState: (value) => {
-            set((state) => ({ ...state, ...value }));
-        },
-    }));
-
-export const createCookiesStoreOnServer = (cookieStoreStateFromContext: string | null) => {
-    let initState = getDefaultInitState();
-
-    if (cookieStoreStateFromContext) {
-        initState = JSON.parse(cookieStoreStateFromContext);
-    }
-
-    return createStore<CookiesStore>()((set) => ({
-        ...initState,
-        setCookiesStoreState: (value) => {
-            set((state) => ({ ...state, ...value }));
-        },
-    }));
-};
 
 export const useCookiesStore = <T>(selector: (store: CookiesStore) => T): T => {
     const cookiesStoreContext = useContext(CookiesStoreContext);
